+++ docs/src/whats-new.md
@@ -2,6 +2,10 @@
 
 - [`loglayer` Changelog](/core-changelogs/loglayer-changelog)
 
+## Jan 17, 2025
+
+- Added the [Log File Rotation Transport](https://loglayer.dev/transports/log-file-rotation) to write logs to files and rotate them
+
 ## Jan 15, 2025
 
 - Added the [Filter Plugin](https://loglayer.dev/plugins/filter) for filtering log messages.